Add an option to not send the lock password in the SMS lock message.
When the app receives an SMS to lock the device it sends a SMS response
informing the sender the device was locked. The password used to lock
the device is sent along with the confirmation. A checkbox now exists in
the lock settings to enable/disable the sending of the password in the
SMS response.